requirements.txt
numpy>=1.20.0,<2.0.0
pillow>=9.0.0,<10.0.0
onnxruntime-gpu>=1.11.0,<2.0.0
fastapi>=0.87.0,<1.0.0
uvicorn>=0.17.0,<1.0.0
python-multipart>=0.0.5,<0.1.0
boto3>=1.24.0,<2.0.0
sagemaker>=2.100.0,<3.0.0